Italo dev

+++ src/main/java/com/api/bigu/services/CarService.java
@@ -1,6 +1,10 @@
 package com.api.bigu.services;
 
-import com.api.bigu.dto.car.CarDTO;
+import com.api.bigu.dto.car.CarRequest;
+import com.api.bigu.dto.car.CarResponse;
+import com.api.bigu.exceptions.CarNotFoundException;
+import com.api.bigu.exceptions.NoCarsFoundException;
+import com.api.bigu.exceptions.UserNotFoundException;
 import com.api.bigu.models.Car;
 import com.api.bigu.models.User;
 import com.api.bigu.repositories.CarRepository;
@@ -10,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.swing.text.html.Option;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,36 +23,56 @@ import java.util.Optional;
 public class CarService {
 
     @Autowired
+    private CarMapper carMapper;
+
+    @Autowired
     private CarRepository carRepository;
 
     @Autowired
     private UserService userService;
 
-    public Optional<Car> findCarById(Integer carId) {
-        return carRepository.findById(carId);
+    public Optional<Car> findCarById(Integer carId) throws CarNotFoundException {
+        Optional<Car> car;
+
+        if (carRepository.findById(carId).isPresent()) {
+            car = carRepository.findById(carId);
+        }
+        else {
+            throw new CarNotFoundException();
+        }
+
+        return car;
     }
 
-    public void deleteById(Integer carId) {
+    public void deleteById(Integer carId) throws CarNotFoundException {
+        if (!carRepository.existsById(carId)) {
+            throw new CarNotFoundException();
+        }
         carRepository.deleteById(carId);
     }
 
-    @SneakyThrows
-    public void addCarToUser(CarDTO carDTO) {
-        Car car = carDTO.toEntity();
-        car.setUser(userService.findUserById(carDTO.getUserId()));
+    public CarResponse addCarToUser(Integer userId, CarRequest carRequest) throws UserNotFoundException {
+        Car car = carMapper.toCar(userId, carRequest);
+        userService.findUserById(userId).getCars().put(car.getPlate(), car);
         carRepository.save(car);
+        return carMapper.toCarResponse(car);
     }
 
-    @SneakyThrows
-    public List<Car> findCarsByUserId(Integer userId) {
-        return carRepository.findAllByUser(userService.findUserById(userId));
+    public List<CarResponse> findCarsByUserId(Integer userId) throws UserNotFoundException, NoCarsFoundException {
+        List<Car> userCars = userService.findUserById(userId).getCars().values().stream().toList();
+        List<CarResponse> carsResponse = new ArrayList<>();
+        for (Car car: userCars
+             ) {
+            carsResponse.add(carMapper.toCarResponse(car));
+        }
+        return carsResponse;
     }
 
-    @SneakyThrows
-    public void removeCarFromUser(Integer userId, Integer carId) {
+    public void removeCarFromUser(Integer userId, Integer carId) throws UserNotFoundException, CarNotFoundException {
         User user = userService.findUserById(userId);
-        Car car = findCarById(carId).orElseThrow(() -> new Exception("Car not found"));
-        if (car.getUser().equals(user)) {
+        Car car = findCarById(carId).get();
+        if (user.getCars().containsKey(car.getPlate())) {
+            user.getCars().remove(car.getPlate());
             carRepository.delete(car);
         }
     }
